Add demangling and output filtering support
It still outputs standard library code, which I'd like to fix, but that
comes later!

There's a bug in undname in file mode; lambdas aren't supported
correctly. In single-identifier mode,

  `??R<lambda_9366063389c5f42a00a5088cf24e69de>@@QBEXQADIQBD@Z`

gets demangled to

  ```
  <lambda_9366063389c5f42a00a5088cf24e69de>::operator()(
    char * const,
    unsigned int,
    char const * const)const
  ```

(spaces added for clarity), however in file mode it gets demangled to

  ` ?? operator()<lambda_9366063389c5f42a00a5088cf24e69de>@@QBEXQADIQBD@Z`

It's likely I'll switch back to a single-identifier based version for
the future, however this works now for most code.